+++ packages/extension/src/index.ts
@@ -15,3 +15,5 @@ export * from './materials/group';
 export * from './NodeResize';
 export * from './tools/flow-path';
 export * from './tools/auto-layout';
+export * from './bpmn-adapter/xml2json';
+export * from './bpmn-adapter/json2xml';