+++ examples/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/didi/LogicFlow/compare/examples@1.0.5-alpha.1...examples@1.0.5) (2022-01-18)
+
+**Note:** Version bump only for package examples
+
+
+
+
+
+## [1.0.5-alpha.1](https://github.com/didi/LogicFlow/compare/examples@1.0.5-alpha.0...examples@1.0.5-alpha.1) (2022-01-18)
+
+**Note:** Version bump only for package examples
+
+
+
+
+
+## [1.0.5-alpha.0](https://github.com/didi/LogicFlow/compare/examples@1.0.4...examples@1.0.5-alpha.0) (2022-01-18)
+
+
+### Bug Fixes
+
+* typo ([facccdf](https://github.com/didi/LogicFlow/commit/facccdf3b98d44650c75e34a66ed33f5e0d8f75e))
+
+
+
+
+
 ## [1.0.4](https://github.com/didi/LogicFlow/compare/examples@1.0.0-alpha.11...examples@1.0.4) (2022-01-12)